requirements.txt
absl-py==0.7.1
astor==0.7.1
attrs==19.1.0
audioread==2.1.6
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.0
blis==0.7.4
cachetools==3.1.0
catalogue==2.0.1
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
cymem==2.0.5
decorator==4.4.0
defusedxml==0.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
future==0.18.2
gast==0.2.2
gensim==3.8.3
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
grpcio==1.19.0
h5py==2.9.0
httplib2==0.12.1
idna==2.8
importlib-metadata==3.7.2
instagram-private-api==1.6.0.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lesscpy==0.14.0
librosa==0.6.3
llvmlite==0.28.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mecab-python3==0.996.1
mistune==0.8.4
mock==2.0.0
mpmath==1.1.0
murmurhash==1.0.5
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.5
notebook==5.7.8
numba==0.43.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.2
parso==0.4.0
pathy==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.14.3
ply==3.11
preshed==3.0.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pydantic==1.7.3
Pygments==2.8.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-twitter==3.5
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
regex==2020.11.13
requests==2.21.0
requests-oauthlib==1.3.0
resampy==0.2.1
retrying==1.3.3
rich==9.13.0
rsa==4.0
scikit-learn==0.24.1
scipy==1.2.1
seaborn==0.11.1
Send2Trash==1.5.0
simpy==3.0.11
six==1.12.0
sklearn==0.0
smart-open==3.0.0
soupsieve==2.2.1
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
sympy==1.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
texthero==1.0.9
thinc==8.0.2
threadpoolctl==2.1.0
tornado==5.1.1
tqdm==4.59.0
traitlets==4.3.2
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
wasabi==0.8.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wordcloud==1.8.1
zipp==3.4.1
